ML_Sandbox/requirements.txt
pycaret
streamlit
pandas
pillow
streamlit-option-menu
streamlit-authenticator
hydralit-components
sqlalchemy
faker